database/requirements.txt
SQLAlchemy
alembic
asyncpg
psycopg2-binary
python-dotenv
passlib[bcrypt]
